requirements.txt
tensorflow_gpu
tensorflow_datasets
numpy
pandas
scikit-image